Organized game backend structure into engine folder, and fixed all warnings

+++ engine/src/gameEngine.cpp
@@ -7,13 +7,13 @@ namespace engine
 {
 
 void GameEngine::addPlayer(Player& player) {
-    if(players.size() < 4) {
+    if(static_cast<int>(players.size()) < 4) {
         players.push_back(player);
     }
 }
 
 void GameEngine::removePlayer(Player& player) {
-    for(int i = 0; i < players.size(); i++) {
+    for(size_t i = 0; i < players.size(); i++) {
         if(player == players[i]) {
             players.erase(players.begin() + i);
             return;
